Bring aggregate and bin back into transform.d.ts
