--- types/spec/aggregate.d.ts
@@ -1,22 +1,0 @@
-export type AggregateOp =
-  | 'argmax'
-  | 'argmin'
-  | 'average'
-  | 'count'
-  | 'distinct'
-  | 'max'
-  | 'mean'
-  | 'median'
-  | 'min'
-  | 'missing'
-  | 'q1'
-  | 'q3'
-  | 'ci0'
-  | 'ci1'
-  | 'stdev'
-  | 'stdevp'
-  | 'sum'
-  | 'valid'
-  | 'values'
-  | 'variance'
-  | 'variancep';
