+++ types/spec/transform.d.ts
@@ -1,6 +1,4 @@
 import {
-  AggregateOp,
-  BaseBin,
   Expr,
   ProductionRule,
   NumericValueRef,
@@ -70,6 +68,28 @@ export interface AggregateTransform {
   cross?: boolean;
   drop?: boolean;
 }
+export type AggregateOp =
+  | 'argmax'
+  | 'argmin'
+  | 'average'
+  | 'count'
+  | 'distinct'
+  | 'max'
+  | 'mean'
+  | 'median'
+  | 'min'
+  | 'missing'
+  | 'q1'
+  | 'q3'
+  | 'ci0'
+  | 'ci1'
+  | 'stdev'
+  | 'stdevp'
+  | 'sum'
+  | 'valid'
+  | 'values'
+  | 'variance'
+  | 'variancep';
 
 export interface BinTransform extends BaseBin {
   type: 'bin';
@@ -77,6 +97,50 @@ export interface BinTransform extends BaseBin {
   field: string;
   as: string[];
   signal?: string;
+}
+export interface BaseBin {
+  /**
+   * The number base to use for automatic bin determination (default is base 10).
+   *
+   * __Default value:__ `10`
+   *
+   */
+  base?: number;
+  /**
+   * An exact step size to use between bins.
+   *
+   * __Note:__ If provided, options such as maxbins will be ignored.
+   */
+  step?: number;
+  /**
+   * An array of allowable step sizes to choose from.
+   * @minItems 1
+   */
+  steps?: number[];
+  /**
+   * A minimum allowable step size (particularly useful for integer values).
+   */
+  minstep?: number;
+  /**
+   * Scale factors indicating allowable subdivisions. The default value is [5, 2], which indicates that for base 10 numbers (the default base), the method may consider dividing bin sizes by 5 and/or 2. For example, for an initial step size of 10, the method can check if bin sizes of 2 (= 10/5), 5 (= 10/2), or 1 (= 10/(5*2)) might also satisfy the given constraints.
+   *
+   * __Default value:__ `[5, 2]`
+   *
+   * @minItems 1
+   */
+  divide?: number[];
+  /**
+   * Maximum number of bins.
+   *
+   * __Default value:__ `6` for `row`, `column` and `shape` channels; `10` for other channels
+   *
+   * @minimum 2
+   */
+  maxbins?: number;
+  /**
+   * If true (the default), attempts to make the bin boundaries use human-friendly boundaries, such as multiples of ten.
+   */
+  nice?: boolean;
 }
 
 export interface CollectTransform {
